rename: v2 컨트롤러 패키지 구분 및 네이밍 컨벤션 적용
